Restructured classes
The App.vue (committed earlier) now holds one Client and one Tictactoe object. I decided that the app shell know the most about the HTML makup variables (what is visible when what is "true").

# TicTacToe
Now the TicTacToe class is responsible for all the game breaking stuff.

# Client
Deleted all references to TicTacToe because the Client shell only be the "communcator" to the server.